+++ db/schema.rb
@@ -15,15 +15,7 @@ ActiveRecord::Schema[8.1].define(version: 2026_03_18_000003) do
   enable_extension "pg_catalog.plpgsql"
   enable_extension "vector"
 
-  create_table "users", force: :cascade do |t|
-    t.datetime "created_at", null: false
-    t.string "email", default: "", null: false
-    t.string "encrypted_password", default: "", null: false
-    t.datetime "remember_created_at"
-    t.datetime "reset_password_sent_at"
-    t.string "reset_password_token"
-    t.datetime "updated_at", null: false
-    t.index ["email"], name: "index_users_on_email", unique: true
-    t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true
-  end
+# Could not dump table "documents" because of following StandardError
+#   Unknown type 'vector' for column 'embedding'
+
 end